underscore private resample submodules (#2297)
