--- docs/roman/resample/resample_utils.rst
@@ -1,7 +1,0 @@
-Resample Utilities
-==================
-
-.. currentmodule:: romancal.resample.resample_utils
-
-.. automodule:: romancal.resample.resample_utils
-   :members:
